Split timestamp tests for testing both old and new semantics
The folliwing tests:
 * TestDate
 * TestTime
 * TestTimeWithTimeZone
 * TestTimestamp
 * TestTimestampWithTimeZone

Are split into base, and two concrete subclasses for:
 * Testing old timestamp semantics (Legacy suffix)
 * Testing new timestamp semantics (no suffix).

This is just technical change. No tests are changed. New semantics is
disabled for now.
