+++ presto-main/src/test/java/com/facebook/presto/type/TestTimeLegacy.java
@@ -13,14 +13,11 @@
  */
 package com.facebook.presto.type;
 
-import org.testng.annotations.Test;
-
-@Test(enabled = false) // TODO unignore when new semantics is implemented
-public class TestTimestamp
-        extends TestTimestampBase
+public class TestTimeLegacy
+        extends TestTimeBase
 {
-    public TestTimestamp()
+    public TestTimeLegacy()
     {
-        super(false);
+        super(true);
     }
 }
